# core/api/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateApiKeyRequest": {
        "properties": {
          "challenge_id": {
            "description": "The challenge id associated with the previously displayed code",
            "example": "67647f5ecda913e9a2e11b26",
            "type": "string"
          },
          "code": {
            "description": "The 4-digit code retrieved from Anytype Desktop app",
            "example": "1234",
            "type": "string"
          }
        },
        "type": "object"
      },
      "CreateApiKeyResponse": {
        "properties": {
          "api_key": {
            "description": "The api key used to authenticate requests",
            "example": "zhSG/zQRmgADyilWPtgdnfo1qD60oK02/SVgi1GaFt6=",
            "type": "string"
          }
        },
        "type": "object"
      },
      "CreateChallengeRequest": {
        "properties": {
          "app_name": {
            "description": "The name of the app that is requesting the challenge",
            "example": "anytype_mcp",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "code": {
            "example": "bad_request",
            "type": "string"
          },
          "message": {
            "example": "Bad request",
            "type": "string"
          },
          "object": {
            "example": "error",
            "type": "string"
          },
          "status": {
            "example": 400,
            "type": "integer"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerauth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "support@anytype.io",
      "name": "Anytype Support",
      "url": "https://anytype.io/contact"
    },
    "description": "This API enables seamless interaction with Anytype's resources - spaces, objects, properties, types, templates, and beyond.",
    "license": {
      "name": "Any Source Available License 1.0",
      "url": "https://github.com/anyproto/anytype-api/blob/main/LICENSE.md"
    },
    "termsOfService": "https://anytype.io/terms_of_use",
    "title": "Anytype API",
    "version": "2025-11-08"
  },
...
  "paths": {
    "/v1/auth/api_keys": {
      "post": {
        "description": "After receiving a `challenge_id` from the `/v1/auth/challenges` endpoint, the client calls this endpoint to provide the corresponding 4-digit code along with the challenge ID. The endpoint verifies that the challenge solution is correct and, if it is, returns an `api_key`. This endpoint is central to the authentication process, as it validates the user's identity and issues a key that can be used for further interactions with the API.",
        "operationId": "create_api_key",
        "parameters": [
          {
            "description": "The version of the API to use",
            "in": "header",
            "name": "Anytype-Version",
            "required": true,
            "schema": {
              "default": "2025-11-08",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiKeyRequest"
              }
            }
          },
          "description": "The request body containing the challenge ID and code",
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateApiKeyResponse"
                }
              }
            },
            "description": "The API key that can be used in the Authorization header for subsequent requests"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Bad request"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Create API Key",
        "tags": [
          "Auth"
        ]
      }
    },
    "/v1/auth/challenges": {
      "post": {
        "description": "Generates a one-time authentication challenge for granting API access to the user's vault. Upon providing a valid `app_name`, the server issues a unique `challenge_id` and displays a 4-digit code within the Anytype Desktop. The `challenge_id` must then be used with the `/v1/auth/api_keys` endpoint to solve the challenge and retrieve an authentication token. This mechanism ensures that only trusted applications and authorized users gain access.",
        "operationId": "create_auth_challenge",
        "parameters": [
          {
            "description": "The version of the API to use",
            "in": "header",
            "name": "Anytype-Version",
            "required": true,
            "schema": {
              "default": "2025-11-08",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChallengeRequest"
              }
            }
          },
          "description": "The request body containing the app name",
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateChallengeResponse"
                }
              }
            },
            "description": "The challenge ID associated with the started challenge"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Bad request"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Create Challenge",
        "tags": [
          "Auth"
        ]
      }
    },
    "/v1/search": {
      "post": {
        "description": "Executes a global search over all spaces accessible to the authenticated user. The request body must specify the `query` text (currently matching only name and snippet of an object), optional filters on types (e.g., \"page\", \"task\"), and sort directives (default: descending by last modified date). File-layout objects (file, image, video, audio, pdf) are excluded from results by default; to include them, list one of the file type keys (\"file\", \"image\", \"video\", \"audio\") in the `types` field. Pagination is controlled via `offset` and `limit` query parameters to facilitate lazy loading in client UIs. The response returns a unified list of matched objects with their metadata and properties.",
        "operationId": "search_global",
        "parameters": [
          {
            "description": "The version of the API to use",
            "in": "header",
            "name": "Anytype-Version",
            "required": true,
            "schema": {
              "default": "2025-11-08",
              "type": "string"
            }
          },
          {
            "description": "The number of items to skip before starting to collect the result set",
            "in": "query",
            "name": "offset",
            "schema": {
              "default": 0,
              "type": "integer"
            }
          },
          {
            "description": "The number of items to return",
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 100,
              "maximum": 1000,
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "description": "The search parameters used to filter and sort the results",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse-Object"
                }
              }
            },
            "description": "The list of objects matching the search criteria"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerauth": []
          }
        ],
        "summary": "Search objects across all spaces",
        "tags": [
          "Search"
        ]
      }
    },
    "/v1/spaces": {
      "get": {
        "description": "Retrieves a paginated list of all spaces that are accessible by the authenticated user. Each space record contains detailed information such as the space ID, name, icon (derived either from an emoji or image URL), and additional metadata. This endpoint is key to displaying a user's workspaces.\nSupports dynamic filtering via query parameters (e.g., ?name[contains]=project). See FilterCondition enum for available conditions.",
        "operationId": "list_spaces",
        "parameters": [
          {
            "description": "The version of the API to use",
            "in": "header",
            "name": "Anytype-Version",
            "required": true,
            "schema": {
              "default": "2025-11-08",
              "type": "string"
            }
          },
          {
            "description": "The number of items to skip before starting to collect the result set",
            "in": "query",
            "name": "offset",
            "schema": {
              "default": 0,
              "type": "integer"
            }
          },
          {
            "description": "The number of items to return",
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 100,
              "maximum": 1000,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse-Space"
                }
              }
            },
            "description": "The list of spaces accessible by the authenticated user"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerauth": []
          }
        ],
        "summary": "List spaces",
        "tags": [
          "Spaces"
        ]
      },
      "post": {
        "description": "Creates a new space based on a supplied name and description in the JSON request body. The endpoint is subject to rate limiting and automatically applies default configurations such as generating a random icon and initializing the workspace with default settings (for example, a default dashboard or home page). On success, the new space’s full metadata is returned, enabling the client to immediately switch context to the new internal.",
        "operationId": "create_space",
        "parameters": [
          {
            "description": "The version of the API to use",
            "in": "header",
            "name": "Anytype-Version",
            "required": true,
            "schema": {
              "default": "2025-11-08",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSpaceRequest"
              }
            }
          },
          "description": "The space to create",
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceResponse"
                }
              }
            },
            "description": "The created space"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitError"
                }
              }
            },
            "description": "Rate limit exceeded"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerauth": []
          }
        ],
        "summary": "Create space",
        "tags": [
          "Spaces"
        ]
      }
    },
...
    "/v1/spaces/{space_id}/chats/{chat_id}/messages/read": {
      "post": {
        "description": "Marks messages within the given order-id range as read. Pass an empty body to mark the entire chat as read. `type` defaults to \"messages\"; use \"mentions\" to mark unread @-mentions instead.",
        "operationId": "read_chat_messages",
        "parameters": [
          {
            "description": "The version of the API to use",
            "in": "header",
            "name": "Anytype-Version",
            "required": true,
            "schema": {
              "default": "2025-11-08",
              "type": "string"
            }
          },
          {
            "description": "The ID of the space",
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "The ID of the chat object",
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadChatMessagesRequest"
              }
            }
          },
          "description": "Read range parameters"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {}
            },
            "description": "Marked as read"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerauth": []
          }
        ],
        "summary": "Read messages",
        "tags": [
          "Chat"
        ]
      }
    },
...
    "/v1/spaces/{space_id}/chats/{chat_id}/reactions/read": {
      "post": {
        "description": "Marks unread reactions in the chat as seen. Pass an empty body to mark every unread reaction.",
        "operationId": "read_chat_reactions",
        "parameters": [
          {
            "description": "The version of the API to use",
            "in": "header",
            "name": "Anytype-Version",
            "required": true,
            "schema": {
              "default": "2025-11-08",
              "type": "string"
            }
          },
          {
            "description": "The ID of the space",
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "The ID of the chat object",
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadChatReactionsRequest"
              }
            }
          },
          "description": "Order id of the last read reaction"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {}
            },
            "description": "Reactions marked as read"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerauth": []
          }
        ],
        "summary": "Read reactions",
        "tags": [
          "Chat"
        ]
      }
    },
...
  },
  "openapi": "3.1.0",
  "servers": [
    {
      "url": "http://127.0.0.1:31009"
    }
  ]
}